Move externalDocuments config option to another file for discoverability and isolation benefits
